requirements.txt
opencv-python==4.8.1.78
tensorflow==2.16.1
mediapipe==0.10.21
numpy==1.26.4
Pillow==10.0.1
ipywidgets==8.1.1
matplotlib==3.7.2
